backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0
anthropic>=0.20.0
python-dotenv>=1.0.0
httpx>=0.24.0
requests>=2.28.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
flask>=3.0.0
